+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixes
 
 - In `ColocalizationScoresToAssay`: Changed marker pair separator from "-" to "/", to avoid string operation issues due to "-" occurring in marker names.
+- In `DensityScatterPlot`: Fixed a bug where the reported percentage of cells in a gate was incorrect when using multiple gates. 
 
 ### Added